Defensively copy IdeaModule's sourceDirs / generatedSourceDirs (#127)
## Before this PR

We were configuring IDEA source dirs (a `Set<File>`) with the expectation that the set was mutable.
That's not necessarily true since anyone can `setSourceDirs` with an immutable set.

## After this PR

We now create an immutable set of the current sourceDirs / generatedSourceDirs, plus the new conjure directory we intend to add, and set that on the IdeaModule.

<!-- Reference any existing GitHub issues, e.g. 'fixes #000' or 'relevant to #000' -->