Remove optimization that sometimes caused infinite loops in TreeMaps
The BuildServicesRegistry is apparently not thread-safe, so we
need to either move this registration to during Gradle configuration
or use some other caching mechanism (e.g. put a cache object on an
extension of the root project).

I have left ConjureRunnerResource mostly intact with the expectation
that it will be either used or removed in a subsequent fix of the
optimization.