consider 'useJakartaNamespaces' option when setting up dependencies
Previously, ConjurePlugin would apply only "jakarta" dependencies to
java subprojects, but this allows projects that are using Jakarta
dependencies for which resolution has selected a higher version that has
the newer "jakarta" namespace to possibly be inconsistent with the
"useJakartaNamespaces" flag (for instance, if the flag were set to
false, it would generate code that would not work, and possibly
vice-versa).

With this change, by default gradle-conjure will generate a project with
the older "javax" dependencies unless "useJakartaNamespaces" is set to
true, and then it will select the newer "jakarta" dependencies.