remove afterEvaluate in some places
Previously, this was added because wiring up dependencies for some
subprojects requires reading flags from a ConjureExtension, and those
values are not available until after the root conjure project is fully
evaluated.

With this change, those dependencies are added via an `addProvider`
instead which defers execution of the code that adds the dependencies
until the project is actually configured, by which time the extension
values have been populated.