set the jakartaPackages flag via a gradle property
If the gradle property `conjure.java.jakartaPackages` is set to `true`,
then the corresponding flag will be set within the `java` options
container in `ConjureExtension` at the time the extension object is
created. This ensures that we can eagerly read the property value and
configure the extension early in the gradle lifecycle, before any
subprojects are configured, and the value will propagate correctly.

This change removes previously-added `afterEvaluate` calls in some
places.